move test types to common typings package
